requirements.txt
requests==2.32.5
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
lxml==5.2.1
pandas==2.2.2
sqlalchemy==2.0.48
plotly==6.6.0
langchain==1.2.10
ollama==0.6.1
langchain-ollama==1.0.1
langchain_experimental==0.4.1
tabulate==0.10.0